requirements.txt
contourpy==1.3.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.57.0
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
MouseInfo==0.1.3
numpy==2.0.2
packaging==25.0
pandas==2.2.3
pillow==11.2.1
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2025.2
rubicon-objc==0.5.0
seaborn==0.13.2
six==1.17.0
tzdata==2025.2
zipp==3.21.0
